+++ DatadogCore/Tests/Objc/DDUIKitRUMActionsPredicateTests.swift
@@ -50,9 +50,6 @@ class DDUIKitRUMActionsPredicateTests: XCTestCase {
 
 #if canImport(SwiftUI)
     func testGivenDefaultPredicate_whenAskingSwiftUIView_itReturnsAction() {
-        guard #available(iOS 13, tvOS 13, *) else {
-            return
-        }
         // Given
         let predicate = objc_DefaultUIKitRUMActionsPredicate()